# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk sistem quiz online",
        "title": "Quiz Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/quiz/result": {
            "get": {
                "description": "Melihat hasil akhir quiz",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Quiz"
                ],
                "summary": "Get quiz result",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Attempt ID",
                        "name": "attempt_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.QuizResult"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/quiz/start": {
            "post": {
                "description": "Memulai quiz dan mengembalikan soal",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Quiz"
                ],
                "summary": "Start quiz",
                "parameters": [
                    {
                        "description": "Start Quiz",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.StartQuizRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/quiz/submit": {
            "post": {
                "description": "Autosave jawaban per soal",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Quiz"
                ],
                "summary": "Submit answer",
                "parameters": [
                    {
                        "description": "Submit Answer",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SubmitAnswerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.QuizResult": {
            "type": "object",
            "properties": {
                "final_score": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "model.StartQuizRequest": {
            "type": "object",
            "properties": {
                "quiz_id": {
                    "type": "integer"
                }
            }
        },
        "model.SubmitAnswerRequest": {
            "type": "object",
            "properties": {
                "answer_text": {
                    "type": "string"
                },
                "attempt_id": {
                    "type": "integer"
                },
                "question_id": {
                    "type": "integer"
                }
            }
        }
    }
}